+++ src/django_upgrade/fixers/compatibility_imports.py
@@ -1,18 +1,3 @@
-"""
-- Replace compatibility imports for
-    - `django.core.exceptions.EmptyResultSet`
-    - `django.core.exceptions.FieldDoesNotExist`
-    - `django.forms.utils.pretty_name`
-    - `django.forms.boundfield.BoundField`
-  See https://docs.djangoproject.com/en/3.1/releases/3.1/#id1
-
-- Replace `JSONField` imports:
-  See https://docs.djangoproject.com/en/3.1/releases/3.1/#features-deprecated-in-3-1
-
-- Replace `django.utils.functional.lru_cache` with `functools.lru_cache`
-  Undocumented change
-"""
-
 from __future__ import annotations
 
 import ast
@@ -108,6 +93,7 @@ REPLACEMENTS_EXACT = {
     (2, 0): {
         "django.utils.functional": {"lru_cache": "functools"},
         "django.utils.decorators": {"ContextDecorator": "contextlib"},
+        "django.http.cookie": {"SimpleCookie": "http.cookies"},
     },
     (3, 1): {
         "django.contrib.postgres.forms": {