+++ tests/fixers/test_compatibility_imports.py
@@ -330,6 +330,7 @@ class TestCompatibility111:
 def test_all_transformed():
     check_transformed(
         """\
+        from django.http.cookie import SimpleCookie
         from django.forms.forms import pretty_name
         from django.forms.forms import BoundField
         from django.db.models.fields import FieldDoesNotExist
@@ -341,6 +342,7 @@ def test_all_transformed():
         from django.contrib.postgres.forms.jsonb import JSONField
         """,
         """\
+        from http.cookies import SimpleCookie
         from django.forms.utils import pretty_name
         from django.forms.boundfield import BoundField
         from django.core.exceptions import FieldDoesNotExist